+++ platforms/android/app/src/main/assets/www/css/mqueries.css
@@ -283,6 +283,7 @@
     }
      .map {
         width:100% !important;
+        height: 100% !important;
          margin:0px;
          padding:0px;
     }
