+++ platforms/android/app/src/main/assets/www/map.html
@@ -1856,7 +1856,7 @@
         </div>
     </div>
     <div class="step1_swipe"></div>
-    <div id="footer" class="resize" style="bottom:0px !important;">
+    <!-- <div id="footer" class="resize" style="bottom:0px !important;">
         <div class="lighthandle" onclick="updateBottomBar();"><i class="fa fa-chevron-up" id="flicker"
                 style="font-size:22px;" aria-hidden="true"></i></div>
         <div id="route_data" class="routedatabg">
@@ -1933,7 +1933,7 @@
             <div style="height:2px; border-bottom:1px solid #ccc; margin: 4px 0px; width:100%;"></div>
         </div>
         <div id="wpListId" class='bg-change' style="max-height:160px; overflow:auto;"></div>
-    </div>
+    </div> -->
     <div class="rotationclass">
         <div class="centerrotation" onclick="rotatemap()" id="rotatemap" title="Default Rotation">
             <img alt="Default Rotation" src="images/RcenterImg.png" class="RcenterImgclass" id="compassimg">
