--- platforms/android/platform_www/plugins/cordova-plugin-splashscreen/www/splashscreen.js
@@ -1,36 +1,0 @@
-cordova.define("cordova-plugin-splashscreen.SplashScreen", function(require, exports, module) {
-/*
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
-*/
-
-var exec = require('cordova/exec');
-
-var splashscreen = {
-    show: function () {
-        exec(null, null, 'SplashScreen', 'show', []);
-    },
-    hide: function () {
-        exec(null, null, 'SplashScreen', 'hide', []);
-    }
-};
-
-module.exports = splashscreen;
-
-});
