--- platforms/android/wrapper.gradle
@@ -1,1 +1,0 @@
-//This file is intentionally just a comment
