+++ www/css/notifications.css
@@ -292,14 +292,14 @@
          width: 50px;
          height: 50px;
          background-position: 10px;
-         top: 176px;
+         top: 6px;
     }
     #wifibutton{
          width: 50px;
          height: 50px;
          background-position: 10px;
-         top: 176px;
-    }l
+         top: 128px;
+    }
     h6{
     	font-size:24px;
     }
